Added icons for history and bookmarks.
Changed from using URL to using String to represent addresses (throughout the implementation). This was because something like "browser://bookmarks" is not a valid URL (and therefor creates an exception) but is still a page I wish to represent.

Moved history from the WebPageTabContent class to the Tab class. Now the back buttons can include navigating to browser content other than just web pages. This means the History page and the Bookmarks page!

Provided a way for the user to select a item from history and visit it again, by double clicking on one, or by using the arrow keys and pressing ENTER.

Configured the HistoryTabContent to not list visits to BrowserPage's of type HISTORY, because otherwise there would always be one at the top of the history window, since you are viewing it now!

Configured the "Most Visited" Page list (uner "History" under the window menu), to exlude all BrowserPage's and only show Page's.